requirements.txt
# Core dependencies
Pillow==9.5.0  # For image processing and handling images
tabulate==0.9.0  # For displaying data in tables in the admin panel
numpy>=1.24.0  # For numerical operations and image analysis
sqlite3-api>=0.1.0  # For database operations

# Image processing dependencies
opencv-python>=4.7.0  # For computer vision tasks

# Utility dependencies
matplotlib>=3.7.0  # For plotting and visualization
pandas>=2.0.0  # For data manipulation and analysis


